Move custom path option from FAB to menu
