+++ vlc-android/res/menu/activity_option.xml
@@ -79,6 +79,12 @@
         android:title="@string/equalizer" />
     <item
         android:orderInCategory="3"
+        android:id="@+id/ml_menu_custom_dir"
+        android:title="@string/add_custom_path"
+        android:contentDescription="@string/add_custom_path_description"
+        android:visible="false"/>
+    <item
+        android:orderInCategory="3"
         android:title="@string/clear_history"
         android:visible="false"
         android:icon="@android:drawable/ic_menu_close_clear_cancel"
